requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
discord.py==2.3.2
Django==5.0.6
djangorestframework==3.15.1
frozenlist==1.4.1
gunicorn==20.1.0
idna==3.7
multidict==6.0.5
python-decouple==3.8
python-dotenv==1.0.1
requests==2.32.2
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
yarl==1.9.4
